+++ app/forms/conversion/edit_project_form.rb
@@ -10,6 +10,7 @@ class Conversion::EditProjectForm
   attribute :incoming_trust_ukprn
   attribute :advisory_board_date, :date
   attribute :advisory_board_conditions
+  attribute :directive_academy_order, :boolean
 
   validates :establishment_sharepoint_link, presence: true, sharepoint_url: true
   validates :incoming_trust_sharepoint_link, presence: true, sharepoint_url: true
@@ -20,6 +21,8 @@ class Conversion::EditProjectForm
   validates :advisory_board_date, presence: true
   validates :advisory_board_date, date_in_the_past: true
 
+  validates :directive_academy_order, inclusion: {in: [true, false]}
+
   def self.new_from_project(project)
     new(
       project: project,
@@ -27,7 +30,8 @@ class Conversion::EditProjectForm
       incoming_trust_sharepoint_link: project.incoming_trust_sharepoint_link,
       incoming_trust_ukprn: project.incoming_trust_ukprn,
       advisory_board_date: project.advisory_board_date,
-      advisory_board_conditions: project.advisory_board_conditions
+      advisory_board_conditions: project.advisory_board_conditions,
+      directive_academy_order: project.directive_academy_order
     )
   end
 
@@ -46,7 +50,8 @@ class Conversion::EditProjectForm
       incoming_trust_sharepoint_link: incoming_trust_sharepoint_link,
       incoming_trust_ukprn: incoming_trust_ukprn,
       advisory_board_date: advisory_board_date,
-      advisory_board_conditions: advisory_board_conditions
+      advisory_board_conditions: advisory_board_conditions,
+      directive_academy_order: directive_academy_order
     )
 
     project.save!